Refactor shared util functions to util package on test extended folder
